requirements.txt
django~=4.2.6
djangorestframework~=3.14.0
django-rest-passwordreset
psycopg2-binary
pyyaml
ujson
requests
django-filter~=23.3